Improved sleeping mechanism
Try running the following in a terminal, while status-line is running in another window:
`while :; do sleep 0.2; touch /tmp/test; sleep 0.2; rm /tmp/test; done`
You'll see that the old version of status-line stops printing lines. It will only continue printing when you stop the oneliner. This is due to how the older version does not keep track of remaining sleep time. The issue is fixed by the changes in this commit.